38015: move/fix/remove some eager tests
