+++ tests/ttnn/nightly/unit_tests/operations/fused/test_single_core_fused_ops.py
