+++ tests/ttnn/nightly/unit_tests/operations/matmul/test_rs_matmul_1d_gather_in0.py
