+++ tests/ttnn/nightly/unit_tests/operations/reduction/test_min_max.py
