+++ tests/ttnn/nightly/unit_tests/operations/reduction/test_prod_nc.py
