+++ tests/ttnn/nightly/unit_tests/operations/reduction/test_sum.py
